Move calculate function from Scores.calculate to BaseSimilarity.calculate_scores()
